project/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
flake8==6.1.0
idna==3.4
lxml==4.9.3
mccabe==0.7.0
pycodestyle==2.11.0
pyflakes==3.1.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.5
XlsxWriter==3.1.6
